requirements.txt
Flask
gunicorn
langchain
openai
python-dotenv==1.0.0